% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renku_load.R
\name{renku_load}
\alias{renku_load}
\title{Load a Data in Renku a environment}
\usage{
renku_load(file_name, e = globalenv())
}
\arguments{
\item{file_name}{Path to the input file}

\item{e}{Environment to load the data. Default is global environment.}
}
\value{
A matrix of the infile
}
\description{
This function enables data loading withing Renku instance. It first load the the git lfs files to the
current virutal machine and then load the data to session.
}
