% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renku_load.R
\name{renku_save}
\alias{renku_save}
\title{Save a Data in a Renku environment}
\usage{
renku_save(..., file_name, e = globalenv())
}
\arguments{
\item{...}{the names of the objects to be saved (as symbols or character strings).}

\item{file_name}{The name of the file where the data will be saved}

\item{e}{Environment to load the data. Default is global environment.}
}
\value{
A matrix of the infile
}
\description{
This function enables data loading withing Renku instance. It first load the the git lfs files to the
current virutal machine and then load the data to session.
}
